requirement.txt
pygraphviz
pylatex
python-dotenv
pydantic
matplotlib
networkx
openai
anthropic
fastapi
uvicorn
ffmpeg
PyMuPDF
pydub
requests